Cast position and momentum to double in rootevent sampler output. Set z to type double.
Former-commit-id: 0afec89e3a6b0e05a68c19610045d270a9dbefa0

+++ include/BDSOutputROOTEventSampler.hh
@@ -22,11 +22,11 @@ template<class T> class BDSOutputROOTEventSampler: public TObject
 public:
   std::string samplerName; //|| Don't split the header
   
-  int                n;    
+  int                n;
   std::vector<T>     energy;
   std::vector<T>     x;
   std::vector<T>     y;
-  float              z;
+  double             z;
   std::vector<T>     xp;
   std::vector<T>     yp;
   std::vector<T>     zp;
